can: mcp25xxfd: optimize SPI reads of FIFOs in can2.0 mode
To reduce the amount of spi_messages send this patch optimizes the
read of RX fifos so that multiple Fifos are read together.

This only works well in CAN2.0 mode where the overhead of prefetching
8 bytes (even if the actual length is 0) is negligable compared
to the SPI-framework overhead.

As we have been issuing less SPI transfers reading SRAM this
also has reduced the likleyhood of SERR happening - it does not
solve the complete issue as seen above.

Change-Id: I0d46ebd03b7951bdcd0272bd9e6fc436542c5e94
Signed-off-by: Martin Sperl <kernel@martin.sperl.org>
Git-commit: 1a138c60e27b19888649872be3add76a76664e12
Git-repo: https://github.com/msperl/linux-rpi/releases/tag/mcp25xxfd-V8.2
Signed-off-by: Ke Du <quic_kedu@quicinc.com>