can: mcp25xxfd: add prediction of CanFD frames sizes based on history
Allow prediction of can frame sizes based on the last 32 Can Frames
received.

Naive histogram approach hast been taken for now.

The first 16 frames are predicted as 0, but after that the prediction is 6.

It should be possible to take this prediction to use bulk reads for CanFD
as well when we have a prediction of length of 48 or 64.

Change-Id: I99666515e8d8a2d7d77888984574779d7cde41e4
Signed-off-by: Martin Sperl <kernel@martin.sperl.org>
Git-commit: 6eb7bb236314da8b9fe45d49d245c7691e063c1d
Git-repo: https://github.com/msperl/linux-rpi/releases/tag/mcp25xxfd-V8.2
Signed-off-by: Ke Du <quic_kedu@quicinc.com>