can: mcp25xxfd: optimize reception of big CanFD frame reception with BRS
In case of predicted DLC=15 and in case at least 33% of those are
sent via BRS be aggressive and use bulk fifo reads to avoid missing
any frames due to releasing FIFOs not as fast as possible.

This would be of most impact when having a 1MHz Can Bus with
a 8MHz data bitrate.

Change-Id: I2a17c2cc83ccb755607b147ee7818537223da915
Signed-off-by: Martin Sperl <kernel@martin.sperl.org>
Git-commit: a3ba69fdb18e7739acfcc5946938beb7df9b3e7d
Git-repo: https://github.com/msperl/linux-rpi/releases/tag/mcp25xxfd-V8.2
Signed-off-by: Ke Du <quic_kedu@quicinc.com>