can: mcp25xxfd: Add Can transmission support
Add un-optimized Can2.0 and CanFD transmission support.

On a Rpi3 we can saturate the CAN bus at 1MHz transmitting Can2.0
frames with DLC=0 for the number of configured tx-fifos.
Afterwards we need some time to recover the state before we can
fill in the fifos again.

With 7 tx fifos we can: send those 7 frames in 0.33ms and then we
wait for 0.26ms so that is a 56% duty cycle.

With 24 tx fifos this changes to: 1.19ms for 24 frames and then we
wait for 0.52ms so that is a 70% duty cycle.

Change-Id: I7ef55ae970140bddc630fd8c076a7aac2b12ff19
Signed-off-by: Martin Sperl <kernel@martin.sperl.org>
Git-commit: 672018dd669c02de3385f07e746abec3e3919fe5
Git-repo: https://github.com/msperl/linux-rpi/releases/tag/mcp25xxfd-V8.2
Signed-off-by: Ke Du <quic_kedu@quicinc.com>