mhi_bus: core: Fix Makefile for module building
The Makefile attempts to build the individual component files as their own
individual modules, instead of one combined module, which is not intended
since it results in compile errors about the internal functions having not
been exported.  Fix the Makefile to generate one single mhi.ko which ends
up addressing all of the module build errors.

Change-Id: Ia80a72ffe670e36d23b0ca30d7b658f46edb39e9
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Signed-off-by: Hemanth Puranik <hpuranik@codeaurora.org>