qaic: Create char dev
Now that we can fully boot the device, we should start making it usable.
The primary interface to the device will be via a char dev.  Add the
necessary framework to detect when the device is fully booted, and create
the char dev at that point.  The device is only usable when it is fully
booted.  The char dev does nothing useful yet, but we can easily build on
this to provide functionality.

Change-Id: I5f8f281bffac7af91c674d93c2c1a1aa1f30f4ae
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Signed-off-by: Hemanth Puranik <hpuranik@codeaurora.org>