qaic: Implement mem ioctl
The mem ioctl allows userspace to allocate and free buffers to be used to
send and receive data from the device via a DMA Bridge channel.  Buffers
can be requested as one of 3 types - sending data to the device, receiving
data from the device, or a combo buffer that can do both (for some
additional overhead).

Change-Id: I76c632cd106ed9750889bc3fabfcdd2b82324c2b
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Signed-off-by: Hemanth Puranik <hpuranik@codeaurora.org>