qaic: Fix rmmod crash
Adding the reset irq handler broke rmmod.  If CONFIG_DEBUG_SHIRQ is defined
then shared irqs get called as part of free_irq().  This causes issues
since the reset handler will then likely reference a freed pointer.
The easy fix is to not make the IRQ shared.  Even though IRQs are preferred
to be implemented as shared, MSIs are device specific, and the reset MSI
is global to the device, so there is no valid way the MSI could actually
be used in a shared IRQ context.

Change-Id: Ibca3abd7301d269e7248149945c06d302a427030
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Signed-off-by: Hemanth Puranik <hpuranik@codeaurora.org>