qaic: Validate direct reads over PCI against the error value
If there is an error with PCI, such as the link went down, the spec states
that returned values will have all the bits high - ie 0xFFFFFFFF for a
32-bit value.

Its possible we are in the middle of an operation such as queuing a buffer
to the device, when the link went down, and the driver hasn't been notified
yet.  Such operations should guard against "invalid" values from PCI errors.

Add the required handling to plug this hole.

Change-Id: I54d88cd1abd3b21f2c2a6ab6ac360daf9ae9cab4
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Signed-off-by: Hemanth Puranik <hpuranik@codeaurora.org>