qaic: Move control response processing off of tasklet context
It turns out that the MHI xfer callbacks (ie qaic_mhi_dl_xfer_cb()) are
invoked in tasklet context.  Therefore it is incorrect for us to use a
mutex in that context.  We either need to use a spinlock, or move to
process context to handle the callback.

We choose process context because this is control path, which is not
latency critical, and also changing from mutex to spinlock has additional
impacts on how to handle the queueing operations.

Change-Id: Ib83ac4144e3d52aecf655695cc2395b56215f37f
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Signed-off-by: Hemanth Puranik <hpuranik@codeaurora.org>