qaic: Refactor control path MHI message handling for adverse scenarios
Sending and receiving MHI messages is currently not very resilient against
adverse scenarios.  Structures can go out of scope when still needed,
memory can be leaked, etc.

We need to handle the following scenarios:
-The normal case where we send a message and get a response
-The error case where we timeout waiting for a response, and either that
response never comes, or comes late
-The error case where our wait is interrupted, such as if the user process
gets a signal
-The error case where our wait returns early, but the response was not
received, such as if we are attempting to clean up from a device crash

Refactor the relevant code to ensure that races are handled via proper
locking, and the above scenarios are gracefully handled.

Change-Id: If4c9416c90f688b70eaff1c8bab4641292551247
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Signed-off-by: Hemanth Puranik <hpuranik@codeaurora.org>