qaic: Handle link resets
Host initiated link resets do not invoke the hotplug path like suprise
link down does.  Therefore we don't get a pci device removal on host
initiated link resets (ie secondary bus reset).  To be notified of these
notifications, we need to register pci error handlers.

When the link goes down, all state on the device is lost, so we need to
do much of the same handling as the hotplus case to cleanup local data
and notify users.  However, since the device is expected to come back,
we can skip a few steps, like destroying the device context.

Note that this change exposed a few issues in the device firmware that need
to be fixed before this goes active, so leave a TODO in to activate this
code in the near future.  We want to build upon the functionality this
change introduces within the driver, thus we want to merge this now rather
than wait for the FW issues to be resolved.

Change-Id: I222d59e64613d5f8f5db3b2b611dabc7a6002348
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Signed-off-by: Hemanth Puranik <hpuranik@codeaurora.org>