serial: msm_geni_serial: Don't depend on clock freq table to set baud rate
Current uart driver depends on predefined root clock frequency table to
select a suitable root clock frequency and configure the clock divider
to get the required baud rate. Setting a clock frequency from the
statically defined list of frequencies may not be supported by
the clock controller driver for that SE clock.

This change is to query the supported clock frequencies at runtime
and configure the clock divider based on the available
supported frequencies.

Change-Id: Iede604ee8d18528b66d57edac83eb54ae2619f5b
Signed-off-by: Chandana Kishori Chiluveru <cchiluve@codeaurora.org>