qaic: Factor link state into account for mhi powerdown
Mhi powerdown has two flavors - one where it accesses the link to put the
device into the reset state, and one where it does not and only cleans up
local state.  We need to inteligently do both.

We should access the link to notify the device when the module is unloaded
so that the device can do cleanup.  However we should not access the link
if the link is down, because attempting to do so at the wrong time can
cause the link to go down again after having recovered which is just
needless activity.

Sadly, both of these scenarios run through the pci device removal path,
so we need to determine what triggered the device removal path to infer
the link state.  Since module exit runs before pci device removal, we
can differentiate that scenario from the others.

Change-Id: I4efa67cc561edbe2d04ad4530a936f10e428a2fa
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Signed-off-by: Hemanth Puranik <hpuranik@codeaurora.org>