qaic: Fix null pointer dereference in qaic_pci_remove()
qaic_pci_remove() can trigger a null pointer dereference if we blindly
attempt to remove a chardev that didn't exist, such as when
qaic_pci_remove() is called before the control channel probe occurs.  Fix
this by only attempting to remove the chardev if we have one allocated in
the first place.

Change-Id: I91bea9a356c415d932159f20a9577e603597e461
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Signed-off-by: Hemanth Puranik <hpuranik@codeaurora.org>