qaic: Fix resource leak on close()
If the processing for close() is invoked, we need to make sure any
resources are cleaned up, otherwise we'll leak them.  The user should have
cleaned up everything already, but perhaps they have a code issue, or were
killed.  We already notify the device that the user is going away via the
terminate transaction, so the device resources are released.  Since close()
only gets processed when the last reference to the file descriptor is
dropped, there can't be any user activity we need to worry about.  The
only resource that needs to be cleaned up are assigned DBCs, so lets do
the required cleanup for those.

Change-Id: If3522c86f61d572687967ef96d7f44fe1a71a907
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Signed-off-by: Hemanth Puranik <hpuranik@codeaurora.org>