qaic: Fix mem_handle access after free
In release_dbc(), we drop a reference to the mem handle, then check to see
if it was queued.  Its possible dropping the reference freed the mem handle
thus we could be accessing freed memory.  A simple reorder of operations
fixes this.

In dbc_irq_handler() we appear to do something similar where we drop a
reference, and then do a complete_all on the mem handle.  At this point in
time, we are guarenteed to always have another reference on the mem handle
(since free_handle cannot drop the last reference while the mem handle is
queued) so dbc_irq_handler() is not actually incorrect.  However, reordering
the operations has no real impact, and make the code look to be correct,
therefore lets do that.

Change-Id: I915821fa75e9d726e2d44dbbff894864744117c6
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Signed-off-by: Hemanth Puranik <hpuranik@codeaurora.org>