mhi_bus: core: Check state before processing power_down
We cannot process a power_down if the power state is DISABLED.  There is
no valid mhi_ctxt in that case, so attepting to process the power_down
will likely result in a null pointer dereference.  If the power state is
DISABLED, there is nothing to do anyways, so just bail early.

Change-Id: Iaeeeeb1fd64782188cf8a4ba0e730fcb40c1d626
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Signed-off-by: Hemanth Puranik <hpuranik@codeaurora.org>
[pkanojiy: Resolve cherry-pick conflict]
Signed-off-by: Pranjal Ramajor Asha Kanojiya <pkanojiy@codeaurora.org>