qaic: Support fatal error handling
When fatal error happens, system reboots and we land into PBL. PBL sends
a system error. MHI need to be reset in such situations.

While doing that, also move qaic_dev_reset_clean_local_state() a bit up,
so that we do not need to have a prototype for it before it's usage in
reset_mhi_work_func().

Change-Id: I342fe7ba860cdf4f6cb2f28ca07e4684a4b705bc
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Signed-off-by: Pratyush Anand <pratanan@codeaurora.org>