qaic: cleanup dbc buffer save
We create dbc buffer during encode activation step. If activation
command is successful then we need to keep this buffer saved. Currently,
we save this buffer in qaic_manage(), based on resource.status. resource
is a local variable and it's status is populated only during activation
command, and hence it could be stale for some other manage command.
Therefore, best place to save these buffer is decode_activate().

Change-Id: I14e02cb01212a7cbc446003e5288dd77c0ac60ef
Signed-off-by: Pratyush Anand <pratanan@codeaurora.org>