qaic: Handle MHI SYSERR
The device will trigger the MHI SYSERR state if it has encountered a
serious error and requires the MHI link to be reset to recover.  The
device will not preserve any state when this occurs, so we need to notify
clients and cleanup any resources.  When the MHI link is reestablished, we
reinit for normal operations.

Change-Id: Ie1dd0ec9b530205fdaa4f4fdcdbe13c4e8a550f7
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Signed-off-by: Pratyush Anand <pratanan@codeaurora.org>