qaic: Make the user API a unique scope
In some kernel configurations, there is already a "struct sem", therefore
the qaic driver causes a symbol conflict.  The API to qaic should be in a
unique "scope" so that it does not conflict with anything else.  Update
API so that everything has a presix of "qaic" which is intended to be
unique.

Change-Id: Ic7e603fb10e552e99c4c4120e7314d1d3de667b5
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Signed-off-by: Pratyush Anand <pratanan@codeaurora.org>