qaic: Implement trace support
Implement traces for the manage ioctl errors so that the exact root cause
can be identified when there are multiple paths which return the same error
code.  We use trace because some of these calls can be performance critical
and it will be useful to correlate these events with the data path in the
future.

Change-Id: Icfa18b2917ce4794d792b8e1b9c08e4b1ca55aac
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Signed-off-by: Pratyush Anand <pratanan@codeaurora.org>