mhi_bus: core: Wait for ready state after reset
After the device has signaled the end of reset by clearing the reset bit,
it will automatically reinit MHI and the internal device structures.  Once
That is done, the device will signal it has entered the ready state.

Signaling the ready state involves sending an interrupt (MSI) to the host
which might cause IOMMU faults if it occurs at the wrong time.

If the controller is being powered down, and possibly removed, then the
reset flow would only wait for the end of reset.  At which point, the host
and device would start a race.  The host may complete its reset work, and
remove the interrupt handler, which would cause the interrupt to be
disabled in the IOMMU.  If that occurs before the device signals the ready
state, then the IOMMU will fault since it blocked an interrupt.  While
harmless, the fault would appear like a serious issue has occurred so let's
silence it by making sure the device hits the ready state before the host
completes its reset processing.

Change-Id: I784fc572bd4e65fe53e3b4a7c7d928a3a2fb06a2
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
[pkanojiy: Resolve cherry-pick conflict]
Signed-off-by: Pranjal Ramajor Asha Kanojiya <pkanojiy@codeaurora.org>