wil6210: AP should not forward eapol packets
In AP mode, the driver checks each received frame, in case it is
multicast or unicast packet targeted to one of the AP clients, the
driver does the routing in L2 level. EAPOL packet is plaintext frame
and it should not follow the above role.

When AP receives an EAPOL packet that is multicast or targeted to
a client in the AP network, the driver should not forward this packet
immediately in L2. EAPOL packet should be indicated to the network
stack which will check the packet validity and will decide if to forward
the packet to its client or not.

Change-Id: Iec9077b485430fa7fe397247599574143242d15a
Signed-off-by: Ahmad Masri <amasri@codeaurora.org>