msm: kgsl: Stop using memdesc->usermem
It shouldn't be illegal for memory descriptors to have multiple virtual
memory mappings under certain circumstances. If that is the case
tracking a single usermem address for each memdesc no longer makes much
sense. Get rid of the memdesc->usermem member and use an atomic counter
to track mappings instead.

Change-Id: I77aa77644b0e4d83505ea577ec51593e9f51a7c4
Signed-off-by: Zhenlin Lian <lianzhenlin@codeaurora.org>