pci: msm: Allow the use of the GICv3 MBI driver as msi-controller
Two different drivers could be managing the MSIs for the PCIe RC, the
MSM driver, or the GICv3 MBI driver.  The MSM driver assumes it is
always the MSM driver, so init will fail if the MBI driver is used.  Fix
init to only use the MSM driver if that is what the DT has specified.

Change-Id: I3878bff21d4be28c20926f932846d977afacd036
Signed-off-by: Jeffrey Hugo <quic_jhugo@quicinc.com>